Task Controller comments
Added comment handling to task controller.

+++ backend/schedule/src/main/java/tech/project/schedule/services/TaskCommentService.java
@@ -5,10 +5,13 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import tech.project.schedule.exception.ApiException;
+import tech.project.schedule.model.enums.GlobalRole;
 import tech.project.schedule.model.enums.ProjectUserRole;
+import tech.project.schedule.model.project.Project;
 import tech.project.schedule.model.task.Task;
 import tech.project.schedule.model.task.TaskComment;
 import tech.project.schedule.model.user.User;
+import tech.project.schedule.repositories.ProjectRepository;
 import tech.project.schedule.repositories.TaskCommentRepository;
 import tech.project.schedule.repositories.TaskRepository;
 import tech.project.schedule.repositories.UserRepository;
@@ -16,6 +19,7 @@ import tech.project.schedule.services.utils.GetProjectRole;
 import tech.project.schedule.services.utils.PmAndAssigneeCheck;
 
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
@@ -25,6 +29,7 @@ public class TaskCommentService {
     private final TaskCommentRepository taskCommentRepository;
     private final TaskRepository taskRepository;
     private final UserRepository userRepository;
+    private final ProjectRepository projectRepository;
 
     @Transactional
     public TaskComment addComment(UUID taskId, User user, TaskComment comment){
@@ -53,20 +58,55 @@ public class TaskCommentService {
     }
 
     @Transactional
-    public void deleteAllCommentsForTask(UUID taskId) {
+    public void deleteAllCommentsForTask(UUID taskId, User user) {
+        Task task = taskRepository.findById(taskId)
+                .orElseThrow(() -> new ApiException("Task not found.", HttpStatus.NOT_FOUND));
+        if(GetProjectRole.getProjectRole(user,task.getProject()) != ProjectUserRole.PM){
+            throw new ApiException("You are not allowed to delete comments in this task", HttpStatus.FORBIDDEN);
+        }
         taskCommentRepository.deleteAllByTask_Id(taskId);
     }
 
-    public List<TaskComment> getCommentsForTask(UUID taskId) {
+    public List<TaskComment> getCommentsForTask(UUID taskId, User user) {
+        boolean isAdmin = user.getGlobalRole() == GlobalRole.ADMIN;
+        Task task = taskRepository.findById(taskId)
+                .orElseThrow(() -> new ApiException("Task not found", HttpStatus.NOT_FOUND));
+        boolean isInProject = task.getProject().getMembers().containsKey(user.getId());
+        if(!isAdmin||isInProject){
+            throw new ApiException("You are not allowed to view comments", HttpStatus.FORBIDDEN);
+        }
         return taskCommentRepository.findAllByTask_Id(taskId);
     }
 
-    public List<TaskComment> getCommentsByUser(UUID userId) {
-        return taskCommentRepository.findAllByUser_Id(userId);
+    public List<TaskComment> getUserComments(User currUser, User otherUser) {
+        boolean isAdmin = currUser.getGlobalRole() == GlobalRole.ADMIN;
+        if(isAdmin||currUser.getId().equals(otherUser.getId())){return taskCommentRepository.findAllByUser_Id(currUser.getId());}
+
+        List<TaskComment> userComments = new ArrayList<>();
+        List<TaskComment> allComments = taskCommentRepository.findAllByUser_Id(currUser.getId());
+
+        List<Project> allProject = projectRepository.findAll();
+
+        for (Project project : allProject) {
+            for(TaskComment comment : allComments) {
+                if(project.getMembers().containsKey(currUser.getId())
+                        && project.getMembers().containsKey(otherUser.getId())
+                        && comment.getTask().getProject().equals(project)){
+                    userComments.add(comment);
+                }
+            }
+        }
+        return userComments;
     }
 
-    public TaskComment getCommentById(UUID commentId) {
-        return taskCommentRepository.findById(commentId).orElseThrow(() -> new ApiException("Comment not found", HttpStatus.NOT_FOUND));
+    public TaskComment getCommentById(UUID commentId, User user) {
+        TaskComment comment = taskCommentRepository.findById(commentId)
+                .orElseThrow(() -> new ApiException("Comment not found", HttpStatus.NOT_FOUND));
+        if(PmAndAssigneeCheck.checkIfNotPmAndAssignee(comment.getTask().getId(), user)){
+            throw new ApiException("You are not allowed to view this comment", HttpStatus.FORBIDDEN);
+        }
+        return comment;
+
     }
 
 }
